feat(public): replace TagCloud with Explore block on HomePage
- Replace weighted font-size TagCloud with plain-pill ExploreBlock
- Add 'All tags' link and section header
- Update HomePage to mount ExploreBlock
- Add dedicated CSS for ExploreBlock
- Follows public-ux-proposal (add/remove summary)
